+++ src/gui/__init__.py
@@ -1,0 +1,3 @@
+"""
+The gui package contains everything that's directly connected to gui components.
+"""
