+++ src/gui/uielements/__init__.py
@@ -1,0 +1,3 @@
+"""
+In this package the transformed ui -> py files will be located.
+"""
