+++ src/gui/uihandler/__init__.py
@@ -1,0 +1,3 @@
+"""
+In this package the handler for the transformed ui -> py files will be located.
+"""
