+++ src/player/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains all elements connected to the player.
+"""
